bytecode directory + compute_val closure
