+++ jolt-core/src/jolt/vm/bytecode/hamming_weight.rs
