+++ jolt-core/src/jolt/vm/mod.rs
@@ -541,7 +541,6 @@ where
 }
 
 pub mod bytecode;
-pub mod hamming_weight;
 pub mod instruction_lookups;
 pub mod output_check;
 pub mod ram;
